finish Transition_Function

+++ src/tape.cpp
@@ -61,7 +61,8 @@ void Tape::update(char write_character, direction move_direction) {
     // array of characters terminated by a null byte gets implicitly type
     // converted into a string
     throw Crash("Left move from first cell");
-  if ((move_direction == 'R') && (current_cell == cells.length() - 1))
+  if ((move_direction == 'R') &&
+      (current_cell == static_cast<int>(cells.length() - 1)))
     // "bad_alloc" is the type of exception thrown when out of memory from
     // adding cells (derived from the base class "exception") virtual function
     // "what" is implemented in the C++ class "exception"
@@ -76,7 +77,7 @@ void Tape::update(char write_character, direction move_direction) {
 string Tape::left(int maximum_number_of_cells) const {
   int first_cell = max(0, current_cell - maximum_number_of_cells);
   string value = cells.substr(first_cell, current_cell - first_cell);
-  if (value.length() < current_cell)
+  if (static_cast<int>(value.length()) < current_cell)
     value.insert(0, "<");
   return value;
 }
@@ -87,7 +88,7 @@ string Tape::right(int maximum_number_of_cells) const {
     --end_cell;
   int last_cell = min(end_cell, current_cell + maximum_number_of_cells - 1);
   string value = cells.substr(current_cell, last_cell - current_cell + 1);
-  if (value.length() < end_cell - current_cell + 1)
+  if (static_cast<int>(value.length()) < end_cell - current_cell + 1)
     value.append(">");
   return value;
 }
